Add files via upload
This is my concept. I think the code is quite messy and some improvements can be made to the files.
Currently, you can upload an image to a tv, The site will automatically rename the file to the naming convention and Will store it in the correct location.

Im using IIS to host this site I seem to have no issues but when setting up I had to change the PHP upload size and I also had some issues with file permissions.